requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.4
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.0
pycparser==2.22
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
